refactor: unify template dialogs
